MacroConsole: add `macros?` command to list current macros in REPL
The literal command `macros?` now prints a human-readable list of macros
that are currently imported into the REPL session (or says that no macros
are imported, if so).

This shadows the `obj?` docstring lookup syntax for the MacroPy special object
`macros`, but that's likely not needed. That can still be invoked manually,
using `imacropy.doc(macros)`.