Fix tab completion in interactive mode
For how to get `rlcompleter` to complete from a namespace other than `__main__`
(which here, in interactive mode, is `macropy3` itself, not the user session),
see the stdlib source code. The `rlcompleter` module actually publishes this
feature, although the docs don't talk about it.

We simply create a new namespace (an empty dict) and use that as both the
completion namespace and the locals namespace of the REPL.

https://docs.python.org/3/library/rlcompleter.html#module-rlcompleter
https://github.com/python/cpython/blob/3.8/Lib/rlcompleter.py